Add project data
We first set the path to the preprocessed image data and create lists of image filenames and corresponding labels. Then, we use the train_test_split() function from scikit-learn to split the data into training, validation, and test sets based on the specified ratios. Finally, we create separate directories for each split and save the corresponding images and labels to the appropriate directories using symbolic links