Add files via upload
This commit adds a requirements.txt file to the project, which lists all the dependencies needed to run the MelanomaDetection model. The file includes packages such as torch, torchvision, numpy, pandas, matplotlib, scikit-learn, and scipy, among others. This file will make it easier for others to install the necessary packages to run the model on their own systems.